+++ doc/CHANGELOG.md
@@ -42,6 +42,7 @@ There are a couple of linting changes as well:
 
 <br/><br/>
 
+
 ## v2.0.2.0 - 2024-04-04 - [PR#318](https://github.com/NOAA-OWP/ras2fim/pull/318)
 
 During some testing, it was discovered that some feature geocurves csv's were not being created.  We also found some more error conditions that could stop the entire processing run and have addressed those here. We also discovered some inconsistencies of projections at different parts of post processing such as inundation and benchmark tools.